prettier files

+++ contracts/DecodeBlock.sol
@@ -57,13 +57,13 @@ library DecodeBlock {
         uint8 BlockIDFlag;
         address ValidatorAddress;
         uint64 TimestampMs;
-        bytes Signature;//[R || S || V]
+        bytes Signature; //[R || S || V]
     }
 
-    struct voteForSign{
-        SignedMsgType  Type;        
-        uint64 Height;      
-        uint32 Round;       
+    struct voteForSign {
+        SignedMsgType Type;
+        uint64 Height;
+        uint32 Round;
         bytes32 BlockID;
         uint64 TimestampMs;
         string ChainID;
@@ -71,13 +71,13 @@ library DecodeBlock {
 
     uint8 constant BLOCK_FLAG_COMMIT = 2;
 
-    enum SignedMsgType{
+    enum SignedMsgType {
         UnknownType,
         // Votes
-	    PrevoteType,
-	    PrecommitType,
-	    // Proposals
-	    ProposalType
+        PrevoteType,
+        PrecommitType,
+        // Proposals
+        ProposalType
     }
 
     enum HeaderProperty {
@@ -104,78 +104,115 @@ library DecodeBlock {
         Commit
     }
 
-    function verifyHeader(bytes memory headerRlpBytes , bytes memory commitRlpBytes ,address[] memory validators,uint64[] memory votePowers,uint64 votingPowerNeeded)internal pure returns(bool){
-        // ToDo:verify header base data 
+    function verifyHeader(
+        bytes memory headerRlpBytes,
+        bytes memory commitRlpBytes,
+        address[] memory validators,
+        uint64[] memory votePowers,
+        uint64 votingPowerNeeded
+    ) internal pure returns (bool) {
+        // ToDo:verify header base data
         bytes32 hash = msgHash(headerRlpBytes);
         Commit memory commit = decodeCommit(commitRlpBytes.toRlpItem());
-        require(commit.BlockID == hash,"HeaderHash and BlockID are not equal");
-        
+        require(commit.BlockID == hash, "HeaderHash and BlockID are not equal");
+
         // verify all signatures
-        require(verifyAllSignature(commit,validators,votePowers,true,false,votingPowerNeeded),"failed to verify all signature");
+        require(
+            verifyAllSignature(commit, validators, votePowers, true, false, votingPowerNeeded),
+            "failed to verify all signature"
+        );
 
         return true;
     }
 
-    function verifyAllSignature(Commit memory commit,address[] memory validators,uint64[] memory votePowers, bool lookUpByIndex, bool countAllSignatures,uint64 votingPowerNeeded) internal pure returns(bool){
-       require(votePowers.length == validators.length,"incorrect length");
-       uint64 talliedVotingPower;
-       uint256 idx;
-       for (uint256 i = 0;i < commit.Signatures.length;i++){
+    function verifyAllSignature(
+        Commit memory commit,
+        address[] memory validators,
+        uint64[] memory votePowers,
+        bool lookUpByIndex,
+        bool countAllSignatures,
+        uint64 votingPowerNeeded
+    ) internal pure returns (bool) {
+        require(votePowers.length == validators.length, "incorrect length");
+        uint64 talliedVotingPower;
+        uint256 idx;
+        for (uint256 i = 0; i < commit.Signatures.length; i++) {
             address vaddr = commit.Signatures[i].ValidatorAddress;
 
             if (lookUpByIndex) {
-                require(vaddr == validators[i] , "validator not exist");
+                require(vaddr == validators[i], "validator not exist");
                 idx = i;
-		    }else{
+            } else {
                 // ToDo:
             }
 
-            bytes memory signMsg =  voteSignBytes(commit,"evm_3334",i);
-            require(verifySignature(vaddr , signMsg , commit.Signatures[i].Signature),"failed to verify signature");
+            bytes memory signMsg = voteSignBytes(commit, "evm_3334", i);
+            require(verifySignature(vaddr, signMsg, commit.Signatures[i].Signature), "failed to verify signature");
 
             talliedVotingPower += votePowers[idx];
 
             if (!countAllSignatures && talliedVotingPower > votingPowerNeeded) {
-			    return true;
-		    }
-       }
+                return true;
+            }
+        }
 
-       if(talliedVotingPower <= votingPowerNeeded){
-           return false;
-       }
-       return true;
+        if (talliedVotingPower <= votingPowerNeeded) {
+            return false;
+        }
+        return true;
     }
 
-    function verifySignature(address addr,bytes memory signMsg, bytes memory sig) internal pure returns(bool){
+    function verifySignature(
+        address addr,
+        bytes memory signMsg,
+        bytes memory sig
+    ) internal pure returns (bool) {
         bytes32 hash = msgHash(signMsg);
-        (uint8 v , bytes32 r,bytes32 s) = getVRS(sig);
-        address recAddr = ecrecover(hash,v,r,s);
+        (uint8 v, bytes32 r, bytes32 s) = getVRS(sig);
+        address recAddr = ecrecover(hash, v, r, s);
         return (recAddr == addr);
     }
 
-    function recoverSignature(bytes memory signMsg, bytes memory sig) internal pure returns(address){
+    function recoverSignature(bytes memory signMsg, bytes memory sig) internal pure returns (address) {
         bytes32 hash = msgHash(signMsg);
-        (uint8 v , bytes32 r,bytes32 s) = getVRS(sig);
-        address recAddr = ecrecover(hash,v,r,s);
+        (uint8 v, bytes32 r, bytes32 s) = getVRS(sig);
+        address recAddr = ecrecover(hash, v, r, s);
         return recAddr;
     }
 
-    function getVRS(bytes memory sig)internal pure returns(uint8 v, bytes32 r,bytes32 s){
-        require(sig.length == 65,"signature with wrong length");
-        assembly{    
-            v := mload(add(sig,0x41))
-            r := mload(add(sig,0x20))
-            s := mload(add(sig,0x40))
+    function getVRS(bytes memory sig)
+        internal
+        pure
+        returns (
+            uint8 v,
+            bytes32 r,
+            bytes32 s
+        )
+    {
+        require(sig.length == 65, "signature with wrong length");
+        assembly {
+            v := mload(add(sig, 0x41))
+            r := mload(add(sig, 0x20))
+            s := mload(add(sig, 0x40))
         }
-        if (v==0 || v ==1){
-            v+=27;
+        if (v == 0 || v == 1) {
+            v += 27;
         }
-        return (v,r,s);
+        return (v, r, s);
     }
-    function getVRSCalldata(bytes calldata sig)internal pure returns(uint8 v, bytes32 r,bytes32 s){
-        require(sig.length == 65,"signature with wrong length");
-        (v ,r, s) = (uint8(sig[64]),bytes32(sig[:32]),bytes32(sig[32:64]));
-        return (v,r,s);
+
+    function getVRSCalldata(bytes calldata sig)
+        internal
+        pure
+        returns (
+            uint8 v,
+            bytes32 r,
+            bytes32 s
+        )
+    {
+        require(sig.length == 65, "signature with wrong length");
+        (v, r, s) = (uint8(sig[64]), bytes32(sig[:32]), bytes32(sig[32:64]));
+        return (v, r, s);
     }
 
     /*
@@ -185,7 +222,7 @@ library DecodeBlock {
         // TODO:decode bloom
 
         RLPReader.RLPItem[] memory list = decodeToHeaderList(blockRlpBytes);
-        require(list.length == 21,"Incorrect number of header properties" );
+        require(list.length == 21, "Incorrect number of header properties");
         header.hashData = decodeHashData(list);
         header.baseData = decodeBaseData(list);
         header.validatorData = decodeValidatorData(list);
@@ -234,21 +271,21 @@ library DecodeBlock {
         commit.Round = uint32(property(list, 1).toUint());
         commit.BlockID = bytes32(property(list, 2).toUint());
 
-        require(property(list, 3).isList(),"commit.Signatures should be list");
+        require(property(list, 3).isList(), "commit.Signatures should be list");
         RLPReader.RLPItem[] memory csList = property(list, 3).toList();
         commit.Signatures = new CommitSig[](csList.length);
-        for (uint i=0;i<csList.length;i++){
+        for (uint256 i = 0; i < csList.length; i++) {
             commit.Signatures[i] = decodeCommitSig(csList[i]);
         }
     }
 
     function decodeCommitSig(RLPReader.RLPItem memory csItem) internal pure returns (CommitSig memory cs) {
-        require(csItem.isList(),"CommmitSig RLP item should be list");
-        RLPReader.RLPItem[] memory list  = csItem.toList();
-        cs.BlockIDFlag = uint8(property(list,0).toUint());
-        cs.ValidatorAddress = property(list,1).toAddress();
-        cs.TimestampMs = uint64(property(list,2).toUint());
-        cs.Signature = property(list,3).toBytes();
+        require(csItem.isList(), "CommmitSig RLP item should be list");
+        RLPReader.RLPItem[] memory list = csItem.toList();
+        cs.BlockIDFlag = uint8(property(list, 0).toUint());
+        cs.ValidatorAddress = property(list, 1).toAddress();
+        cs.TimestampMs = uint64(property(list, 2).toUint());
+        cs.Signature = property(list, 3).toBytes();
     }
 
     function decodeNextValidators(bytes memory blockRlpBytes) internal pure returns (address[] memory) {
@@ -264,35 +301,38 @@ library DecodeBlock {
         array = _decodeNextValidatorPowers(_NextValidatorPowers);
     }
 
-    function decodeRlp(bytes memory rlp)internal pure returns(bytes[] memory res){
+    function decodeRlp(bytes memory rlp) internal pure returns (bytes[] memory res) {
         RLPReader.RLPItem[] memory list = rlp.toRlpItem().toList();
 
-        res =new bytes[](list.length);
-        for (uint256 i = 0;i<list.length;i++){
+        res = new bytes[](list.length);
+        for (uint256 i = 0; i < list.length; i++) {
             res[i] = RLPReader.toBytes(list[i]);
         }
-        
     }
 
-    function voteSignBytes(Commit memory commit,string memory chainId,uint idx)internal pure returns(bytes memory){
+    function voteSignBytes(
+        Commit memory commit,
+        string memory chainId,
+        uint256 idx
+    ) internal pure returns (bytes memory) {
         voteForSign memory vfs;
         vfs.Type = SignedMsgType.PrecommitType;
         vfs.Height = commit.Height;
         vfs.Round = commit.Round;
-        if (commit.Signatures[idx].BlockIDFlag == BLOCK_FLAG_COMMIT){
+        if (commit.Signatures[idx].BlockIDFlag == BLOCK_FLAG_COMMIT) {
             vfs.BlockID = commit.BlockID;
         }
 
         vfs.TimestampMs = commit.Signatures[idx].TimestampMs;
-        vfs.ChainID = chainId; 
+        vfs.ChainID = chainId;
         return encodeToRlpBytes(vfs);
     }
 
-    function headerHash(bytes memory blockRlpBytes) internal pure returns(bytes32){
+    function headerHash(bytes memory blockRlpBytes) internal pure returns (bytes32) {
         // TODO
     }
 
-    function encodeToRlpBytes(voteForSign memory vfs) internal pure returns(bytes memory){
+    function encodeToRlpBytes(voteForSign memory vfs) internal pure returns (bytes memory) {
         bytes[] memory List = new bytes[](6);
         List[0] = RLPEncode.encodeUint(uint256(vfs.Type));
         List[1] = RLPEncode.encodeUint(uint256(vfs.Height));
@@ -304,7 +344,7 @@ library DecodeBlock {
         return RLPEncode.encodeList(List);
     }
 
-    function msgHash(bytes memory signMsg) internal pure returns(bytes32) {
+    function msgHash(bytes memory signMsg) internal pure returns (bytes32) {
         return signMsg.toRlpItem().rlpBytesKeccak256();
     }
 
@@ -319,8 +359,4 @@ library DecodeBlock {
     function _decodeNextValidatorPowers(RLPReader.RLPItem memory item) private pure returns (uint64[] memory array) {
         array = item.toUint64Array();
     }
-
-    // function arrayToMap(bytes[] memory array)private pure returns( mapping(bytes32=>uint8)  memory){
-
-    // }
 }
